+++ CHANGELOG.md
@@ -7,8 +7,10 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 3.0.0
 ------
 
-**CHANGES**
+**ENHANCEMENTS**
+- Support restart/reboot for instance type with instance store (ephemeral drives).
 
+**CHANGES**
 - Drop support for SGE and Torque schedulers.
 - Drop support for CentOS8.
 - Remove nodewatcher, sqswatcher, jobwatcher related code.
@@ -21,8 +23,6 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Make PATH include required directories for every user and recipes context.
 - Fail cluster creation when IMDS lockdown is not working correctly.
 - Make sudoers secure_path include the same directories in every platform.
-- Support restart/reboot for instance with instance store (ephemeral drives).
-- Support restart/reboot for instance with instance store (ephemeral drives).
 - Remove option for instance store software encryption (encrypted_ephemeral).  
 - Add support for iptables restore on instance reboot.
 - Allow IMDS access for dcv user when dcv is enabled.